+++ noodles-bam/src/async/writer.rs
@@ -1,7 +1,3 @@
-mod builder;
-
-pub use self::builder::Builder;
-
 use std::ffi::CString;
 
 use noodles_bgzf as bgzf;
@@ -196,19 +192,6 @@ impl<W> Writer<bgzf::AsyncWriter<W>>
 where
     W: AsyncWrite + Unpin,
 {
-    /// Creates an async BAM writer builder.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use noodles_bam as bam;
-    /// let builder = bam::AsyncWriter::builder(Vec::new());
-    /// let writer = builder.build();
-    /// ```
-    pub fn builder(inner: W) -> Builder<W> {
-        Builder::new(inner)
-    }
-
     /// Creates an async BAM writer with a default compression level.
     ///
     /// # Examples